+++ commands/auth.py
@@ -3,10 +3,12 @@ import argparse
 import sys
 from typing import List, Optional, Callable
 from sptfy import Sptfy
-from helpers import get_command_usage, login_required, get_cache_file_path
+from helpers import get_command_usage, login_required, get_cache_file_path, time_taken
+from log import logger
 
 
 def auth_command(original_args: List[str], sptfy: Sptfy) -> None:
+    logger.debug("'auth' command invoked")
     if len(original_args) < 1:
         print(get_command_usage(
             command=AUTH_COMMAND_NAME,
@@ -17,6 +19,7 @@ def auth_command(original_args: List[str], sptfy: Sptfy) -> None:
     subcommand = original_args[0]
 
     if subcommand == "help":
+        logger.debug("'auth' help command invoked")
         print(get_command_usage(
             command=AUTH_COMMAND_NAME,
             subcommands=AUTH_COMMAND_SUBCOMMANDS
@@ -25,6 +28,7 @@ def auth_command(original_args: List[str], sptfy: Sptfy) -> None:
 
     subcommand_function: Optional[Callable] = AUTH_COMMAND_SUBCOMMANDS.get(subcommand, None)
     if subcommand_function is None:
+        logger.debug(f"'auth' '{subcommand}' subcommand not found")
         print(f"Unknown subcommand '{subcommand}', {get_command_usage(
             command=AUTH_COMMAND_NAME, 
             subcommands=AUTH_COMMAND_SUBCOMMANDS
@@ -37,13 +41,17 @@ def auth_command(original_args: List[str], sptfy: Sptfy) -> None:
     )
 
 
+@time_taken
 def login(args: List[str], sptfy: Sptfy) -> None:
+    logger.debug(f"'auth' 'login' subcommand invoked")
     cache_filepath = get_cache_file_path()
 
     login_result = sptfy.auth()
     if login_result:
         print("Login successful!")
+        logger.debug(f"Login success")
     else:
+        logger.debug(f"Login error")
         print("ERROR: Problems logging in.")
         if os.path.exists(cache_filepath):
             print(f"Please delete {cache_filepath} and try again.")
@@ -56,19 +64,23 @@ def logout(args: List[str], sptfy: Sptfy) -> None:
     cache_filepath = get_cache_file_path()
     os.remove(cache_filepath)
     print("Logout successful!")
+    logger.debug(f"Deleted f{cache_filepath} to logout")
 
 
 @login_required
 def current_user(args: List[str], sptfy: Sptfy) -> None:
+    logger.debug(f"Retrieving current user public info")
     parser = get_current_user_parser()
     current_user_args = parser.parse_args(args)
 
     user = sptfy.get_current_user_info()
-
+    logger.debug(f"Showing user name")
     print(f"{user.name}", end="")
     if current_user_args.show_id:
+        logger.debug(f"Showing user Id")
         print(f"\t{user.id}", end="")
     if current_user_args.show_url:
+        logger.debug(f"Showing user url")
         print(f"\t{user.url}", end="")