+++ commands/user_top.py
@@ -2,11 +2,13 @@ import argparse
 import sys
 from typing import List, Optional, Callable
 from sptfy import Sptfy
-from helpers import get_longest_string, get_command_usage, login_required
+from helpers import get_longest_string, get_command_usage, login_required, time_taken
+from log import logger
 
 
 @login_required
 def user_top_command(original_args: List[str], sptfy: Sptfy) -> None:
+    logger.debug(f"'user-top' command invoked")
     if len(original_args) < 1:
         print(get_command_usage(
             command=USER_TOP_COMMAND_NAME,
@@ -16,6 +18,7 @@ def user_top_command(original_args: List[str], sptfy: Sptfy) -> None:
 
     subcommand = original_args[0]
     if subcommand == "help":
+        logger.debug("'user-top' help command invoked")
         print(get_command_usage(
             command=USER_TOP_COMMAND_NAME,
             subcommands=USER_TOP_COMMAND_SUBCOMMANDS
@@ -24,6 +27,7 @@ def user_top_command(original_args: List[str], sptfy: Sptfy) -> None:
 
     subcommand_function: Optional[Callable] = USER_TOP_COMMAND_SUBCOMMANDS.get(subcommand, None)
     if subcommand_function is None:
+        logger.debug(f"'user-top' '{subcommand}' subcommand not found")
         print(f"Unknown subcommand '{subcommand}', {get_command_usage(
             command=USER_TOP_COMMAND_NAME,
             subcommands=USER_TOP_COMMAND_SUBCOMMANDS
@@ -36,15 +40,22 @@ def user_top_command(original_args: List[str], sptfy: Sptfy) -> None:
     )
 
 
+@time_taken
 def top_artists_subcommand(args: List[str], sptfy: Sptfy) -> None:
+    logger.debug(f"Retrieving top artists")
     parser = get_user_top_parser()
     top_artists_args = parser.parse_args(args)
 
     arg_errors = get_user_top_args_errors(top_artists_args)
     if len(arg_errors) > 0:
+        logger.debug(f"{len(arg_errors)} errors found, exiting program")
         for err in arg_errors:
             print(err)
         sys.exit(1)
+
+    logger.debug(f"{len(arg_errors)} errors found, continuing")
+    logger.debug(f"Getting user top artists with input: limit {top_artists_args.limit}, "
+                 f"offset {top_artists_args.offset}, time_range {top_artists_args.time_range}")
 
     artists = sptfy.get_user_top_artists(
         limit=top_artists_args.limit,
@@ -52,20 +63,28 @@ def top_artists_subcommand(args: List[str], sptfy: Sptfy) -> None:
         time_range=top_artists_args.time_range
     )
 
+    logger.debug(f"Showing artists")
     longest_artist_name = get_longest_string([artist.name for artist in artists])
     for artist in artists:
         print(f"{artist.name:<{longest_artist_name}}")
 
 
+@time_taken
 def top_tracks_subcommand(args: List[str], sptfy: Sptfy) -> None:
+    logger.debug(f"Retrieving top tracks")
     parser = get_user_top_parser()
     top_tracks_args = parser.parse_args(args)
 
     arg_errors = get_user_top_args_errors(top_tracks_args)
     if len(arg_errors) > 0:
+        logger.debug(f"{len(arg_errors)} errors found, exiting program")
         for err in arg_errors:
             print(err)
         sys.exit(1)
+
+    logger.debug(f"{len(arg_errors)} errors found, continuing")
+    logger.debug(f"Getting user top tracks with input: limit {top_tracks_args.limit}, "
+                 f"offset {top_tracks_args.offset}, time_range {top_tracks_args.time_range}")
 
     songs = sptfy.get_user_top_tracks(
         limit=top_tracks_args.limit,
@@ -73,6 +92,7 @@ def top_tracks_subcommand(args: List[str], sptfy: Sptfy) -> None:
         time_range=top_tracks_args.time_range
     )
 
+    logger.debug(f"Showing songs and artists")
     longest_song_name = get_longest_string([song.name for song in songs])
     longest_artists_name = get_longest_string([','.join(song.artists) for song in songs])
     for song in songs:
@@ -113,6 +133,7 @@ def get_user_top_parser() -> argparse.ArgumentParser:
 
 
 def get_user_top_args_errors(args: argparse.Namespace) -> List[str]:
+    logger.debug(f"Validating arguments passed")
     errors = []
 
     # Validate limit