+++ helpers/__init__.py
@@ -1,12 +1,13 @@
-import json
 import os
 import getpass
 import sys
 import __main__
+import time
 from pathlib import Path
 from dotenv import load_dotenv
 from typing import Dict, List, Callable
 from dataclasses import dataclass
+from log import logger
 
 
 SPOTIFY_CLIENT_ID_ENV_VARIABLE_STR = "SPOTIFY_CLIENT_ID"
@@ -39,6 +40,7 @@ def null_or_empty(response: str, env_var_name) -> bool:
 
 
 def get_required_environment_variables_as_input() -> EnvironmentVariables:
+    logger.debug(f"Retrieving required environment variables as input")
     spotify_client_id = getpass.getpass(f"{SPOTIFY_CLIENT_ID_ENV_VARIABLE_STR}: ").strip()
     if null_or_empty(spotify_client_id, SPOTIFY_CLIENT_ID_ENV_VARIABLE_STR):
         sys.exit(1)
@@ -61,6 +63,7 @@ def get_required_environment_variables_as_input() -> EnvironmentVariables:
 
 
 def get_required_environment_variables() -> EnvironmentVariables:
+    logger.debug(f"Retrieving required environment variables")
     load_dotenv()
 
     spotify_client_id = os.getenv(SPOTIFY_CLIENT_ID_ENV_VARIABLE_STR, default=None)
@@ -119,8 +122,10 @@ def login_required(func) -> Callable:
         cache_filepath = get_cache_file_path()
 
         if os.path.exists(cache_filepath):
+            logger.debug(f"{cache_filepath} exists and going to function")
             return func(*args, **kwargs)
         else:
+            logger.debug(f"{cache_filepath} does not exist")
             print("Not currently logged in!")
             print("Please log in with: `spotilist auth login`")
             sys.exit(1)
@@ -148,3 +153,14 @@ def get_parent_dir() -> Path:
         main_script_path = Path(os.path.realpath(__main__.__file__))
         parent_dir = main_script_path.parent.absolute()
     return parent_dir
+
+
+def time_taken(func) -> Callable:
+    def wrapper(*args, **kwargs) -> None:
+        start = time.time()
+        func(*args, **kwargs)
+        end = time.time()
+        duration = end - start
+        logger.info(f"Time taken: {round(duration, 2)} seconds")
+
+    return wrapper